+++ .github/CHANGELOG.md
@@ -51,7 +51,8 @@
 - move Docs to UI
 - fill Docs with content
 - LoadData: modernized & minify design
-- minified `any` type and `!` usage
+- minified `any` type and `!` usage; remove Any<T> & F<T> types
+- update document title to current project
 
 ## 0.22.1 (11 Aug 2026)